Face Reconstruction project (IPCV, 3rd quartile)
Scripts & data for Image Processing and Computer Vision project from the previous quartile: 3D face reconstruction. 
In short, we had images of a face taken simultaneously by 3 cameras. We took pairs of images (middle+left, middle+right) and computed disparity map and point cloud out of each of them. Then, we were merging the two point clouds with ICP. For some subjects it worked better, for some worse (especially for the girl, due to her curly hair), and it was ICP stage which seemed to introduce the most mess. Lowering the parameter of inlier ratio in ICP  helped a bit to successfully merge the point clouds for some cases.